requirements.txt
django~=4.0.2
django-heroku
gunicorn
djangorestframework~=3.13.1